requirements.txt
streamlit
yfinance
pandas
numpy
scikit-learn
plotly
matplotlib
joblib
tensorflow-cpu